+++ LoggingLib/Ilog.cs
@@ -6,8 +6,8 @@
         void WriteCounter(string message);
         void WriteLine(string line);
         void ClearCurrentConsoleLine();
-        void ProgressSearch(double counter, double total, string message, string tailMessage = "");
-        void ProgressUpdate(double counter, double total, string message, string tailMessage = "");
+        void ProgressSearch(int counter, int total, string message, string tailMessage = "");
+        void ProgressUpdate(int counter, int total, string message, string tailMessage = "");
         void StatusReport(string message, bool forceNewLine, bool pause = false);
     }
 }
